Change some storage engine signatures and move logic from firebase storage engine to abstract class
